Refactor surface
