+++ src_c/include/palette.h
