+++ src_c/pixelarray.c
@@ -27,7 +27,7 @@
 
 #include "doc/pixelarray_doc.h"
 
-#include "surface.h"
+#include "surface/surface.h"
 #if !defined(BUILD_STATIC)
 static char FormatUint8[] = "B";
 static char FormatUint16[] = "=H";
