+++ src_c/static.c
@@ -365,9 +365,9 @@ PyInit_pygame_static()
 #undef pgSurface_Type
 #undef pgSurface_SetSurface
 
-#include "surface.c"
-#include "simd_blitters_avx2.c"
-#include "simd_blitters_sse2.c"
+#include "surface/surface.c"
+#include "surface/simd_blitters_avx2.c"
+#include "surface/simd_blitters_sse2.c"
 
 #include "window.c"
