+++ src_c/surface/_surface.h
@@ -24,7 +24,7 @@
 #ifndef _SURFACE_H
 #define _SURFACE_H
 
-#include "_pygame.h"
+#include "../_pygame.h"
 #include "surface.h"
 
 #endif
