+++ src_c/surface/alphablit.c
@@ -24,7 +24,7 @@
 
 #define NO_PYGAME_C_API
 #include "_surface.h"
-#include "simd_shared.h"
+#include "../include/simd_shared.h"
 #include "simd_blitters.h"
 
 static void
