+++ src_c/surface/simd_blitters.h
