+++ src_c/surface/simd_fill.h
