+++ src_c/surface/simd_surface_fill_avx2.c
