+++ src_c/surface/surface.c
@@ -28,8 +28,8 @@
 #include "palette.h"
 
 #include "structmember.h"
-#include "pgcompat.h"
-#include "doc/surface_doc.h"
+#include "../pgcompat.h"
+#include "surface_doc.h"
 
 /* stdint.h is missing from some versions of MSVC. */
 #ifdef _MSC_VER
